VULCAN-19 Rename vulcan-liferay-api to vulcan-liferay-portal
