--- modules/apps/foundation/vulcan/vulcan-liferay-api/bnd.bnd
@@ -1,7 +1,0 @@
-Bundle-Name: Liferay Vulcan Liferay API
-Bundle-SymbolicName: com.liferay.vulcan.liferay.api
-Bundle-Version: 1.0.0
-Export-Package:\
-	com.liferay.vulcan.liferay.context,\
-	com.liferay.vulcan.liferay.filter,\
-	com.liferay.vulcan.liferay.identifier
