+++ modules/apps/foundation/vulcan/vulcan-liferay-portal/bnd.bnd
@@ -1,0 +1,7 @@
+Bundle-Name: Liferay Vulcan Liferay Portal
+Bundle-SymbolicName: com.liferay.vulcan.liferay.portal
+Bundle-Version: 1.0.0
+Export-Package:\
+	com.liferay.vulcan.liferay.portal.context,\
+	com.liferay.vulcan.liferay.portal.filter,\
+	com.liferay.vulcan.liferay.portal.identifier
