+++ modules/apps/foundation/vulcan/vulcan-liferay-portal/build.gradle
