+++ modules/apps/foundation/vulcan/vulcan-sample-rest/build.gradle
@@ -6,5 +6,5 @@ dependencies {
 	provided group: "javax.ws.rs", name: "javax.ws.rs-api", version: "2.0.1"
 	provided group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
 	provided project(":apps:foundation:vulcan:vulcan-api")
-	provided project(":apps:foundation:vulcan:vulcan-liferay-api")
+	provided project(":apps:foundation:vulcan:vulcan-liferay-portal")
 }
