+++ modules/apps/foundation/vulcan/vulcan-sample-rest/src/main/java/com/liferay/vulcan/sample/rest/internal/rating/AggregateRatingServiceImpl.java
@@ -16,7 +16,7 @@ package com.liferay.vulcan.sample.rest.internal.rating;
 
 import com.liferay.ratings.kernel.model.RatingsEntry;
 import com.liferay.ratings.kernel.service.RatingsEntryLocalService;
-import com.liferay.vulcan.liferay.identifier.ClassNameClassPKIdentifier;
+import com.liferay.vulcan.liferay.portal.identifier.ClassNameClassPKIdentifier;
 import com.liferay.vulcan.sample.rest.rating.AggregateRating;
 import com.liferay.vulcan.sample.rest.rating.AggregateRatingService;
