+++ modules/apps/foundation/vulcan/vulcan-sample-rest/src/main/java/com/liferay/vulcan/sample/rest/internal/resource/BlogPostingResource.java
@@ -28,9 +28,9 @@ import com.liferay.portal.kernel.service.GroupLocalService;
 import com.liferay.portal.kernel.service.UserService;
 import com.liferay.portal.kernel.util.DateUtil;
 import com.liferay.vulcan.filter.QueryParamFilterType;
-import com.liferay.vulcan.liferay.filter.ClassNameClassPKFilter;
-import com.liferay.vulcan.liferay.filter.GroupIdFilter;
-import com.liferay.vulcan.liferay.identifier.ClassNameClassPKIdentifier;
+import com.liferay.vulcan.liferay.portal.filter.ClassNameClassPKFilter;
+import com.liferay.vulcan.liferay.portal.filter.GroupIdFilter;
+import com.liferay.vulcan.liferay.portal.identifier.ClassNameClassPKIdentifier;
 import com.liferay.vulcan.pagination.PageItems;
 import com.liferay.vulcan.pagination.Pagination;
 import com.liferay.vulcan.resource.Resource;
