+++ modules/apps/foundation/vulcan/vulcan-sample-rest/src/main/java/com/liferay/vulcan/sample/rest/internal/resource/CommentResource.java
@@ -28,8 +28,8 @@ import com.liferay.portal.kernel.security.auth.PrincipalException;
 import com.liferay.portal.kernel.service.IdentityServiceContextFunction;
 import com.liferay.portal.kernel.service.ServiceContext;
 import com.liferay.portal.kernel.service.UserLocalService;
-import com.liferay.vulcan.liferay.context.CurrentUser;
-import com.liferay.vulcan.liferay.filter.ClassNameClassPKFilter;
+import com.liferay.vulcan.liferay.portal.context.CurrentUser;
+import com.liferay.vulcan.liferay.portal.filter.ClassNameClassPKFilter;
 import com.liferay.vulcan.pagination.PageItems;
 import com.liferay.vulcan.pagination.Pagination;
 import com.liferay.vulcan.resource.Resource;
